homework-4 task 1 full complited
