--- test.json
@@ -1,12 +1,0 @@
-{
-    "name": "Вася",
-    "surname": "Петров",
-    "age": 35,
-    "isAdmin": false,
-    "phoneNumbers": [
-      "495 123-1234",
-      "916 123-4567"
-    ]
-  }
-  
-
